rename(menu): dto 디렉토리 분리
